Convert the pi bridge to narrow-grammar delta emission
Pi's translator shrinks to dialect parsing (delta-translation.ts): schema
narrowing, bash/edit/write classification, placeholder stripping, the
ignored-event set, visibility-driven unhandled deltas, and the model
context-window catalog. It is stateless — the per-session turn-state
registry, scoped-item-id factories, accepted-input queue, snapshot diffing,
cumulative-token accumulation, and entropy id minting all moved into the
runtime assembler. Bridge lifecycle sites now speak deltas too: interrupt
emits session.ended instead of a hand-built turn/completed, prompt-settled
and agent_end emit claimIfIdle boundaries, turn/start and steer emit
input.accepted, and session errors ride a settling provider.error.

Bridge and conformance tests assemble the captured thread/delta
notifications through a real delta assembler (the runtime adapter's exact
translation), so the canonical protocol suite still passes end to end.
