+++ packages/agent-runtime/src/pi/bridge/__tests__/model-list.test.ts
@@ -13,7 +13,7 @@ vi.mock("@earendil-works/pi-ai", () => ({
   getSupportedThinkingLevels,
 }));
 
-import { createPiModelContextWindowResolverFrom } from "../../event-translation.js";
+import { createPiModelContextWindowResolverFrom } from "../../delta-translation.js";
 import {
   listPiBridgeModels,
   resetPiModelNetworkRefreshForTests,
